stock-backend/requirements.txt
fastapi
uvicorn
pandas
numpy
yfinance
scikit-learn
mysql-connector-python
sqlalchemy
python-dotenv
python-multipart